move the update functions in a dedicated updates module
