+++ tests/test_continuous.py
@@ -8,13 +8,13 @@ from jax.lax import scan
 from jax.tree_util import Partial
 
 from pyhgf import load_data
-from pyhgf.continuous import (
+from pyhgf.structure import beliefs_propagation
+from pyhgf.typing import Indexes
+from pyhgf.updates.continuous import (
     continuous_input_update,
     continuous_node_update,
     gaussian_surprise,
 )
-from pyhgf.structure import beliefs_propagation
-from pyhgf.typing import Indexes
 
 
 class Testcontinuous(TestCase):
