+++ tests/test_structure.py
@@ -5,9 +5,9 @@ from unittest import TestCase
 
 import jax.numpy as jnp
 
-from pyhgf.continuous import continuous_input_update, continuous_node_update
 from pyhgf.structure import beliefs_propagation, list_branches, trim_sequence
 from pyhgf.typing import Indexes
+from pyhgf.updates.continuous import continuous_input_update, continuous_node_update
 
 
 class TestStructure(TestCase):
